Added uaf thermocouple

+++ far-mars-onboard-firmware/src/daq_send.c
@@ -16,7 +16,7 @@
 
 struct daqSensors {
     uint32_t timestamp;
-    int32_t TC_uaf;
+    int16_t TC_uaf;
     uint8_t HALL_methane, HALL_LOX;
     uint16_t BATT_voltage;
     int16_t PT_methane, PT_LOX, PT_helium, PT_chamber, PT_heliumReg;
